+++ lib/bm25.dart
@@ -1,6 +1,7 @@
 library bm25;
 
 export 'src/bm25_document.dart';
-export 'src/bm25_core.dart';
+export 'src/bm25.dart';
+export 'src/partitioned_bm25.dart';
 export 'src/search_result.dart';
 export 'src/extensions.dart';